Refactor to resolve import loop in tests
